Various bug fixes for semantic and syntax tests. (#264)
* Various bug fixes for semantic and syntax tests.

* Relaxed an assertion that was too strict and updated compilation tests

+++ test/libsolidity/semanticTests/variables/public_state_overridding_mapping_to_dynamic_struct.sol
@@ -1,10 +1,10 @@
 pragma abicoder               v2;
 
-struct S { uint256 v; string s; }
+struct S { uint v; string s; }
 
 contract A
 {
-	function test(uint256 x) external virtual returns (uint256 v, string memory s)
+	function test(uint x) external virtual returns (uint v, string memory s)
 	{
 	    v = x;
 	    s = "test";
@@ -12,15 +12,15 @@ contract A
 }
 contract X is A
 {
-	mapping(uint256 => S) public override test;
+	mapping(uint => S) public override test;
 
 	function set() public { test[42].v = 2; test[42].s = "statevar"; }
 }
 
 
 // ----
-// test(uint256): 0 -> 0, 64, 0
-// test(uint256): 42 -> 0, 64, 0
+// test(uint): 0 -> 0, "" 
+// test(uint): 42 -> 0, ""
 // set() ->
-// test(uint256): 0 -> 0, 64, 0
-// test(uint256): 42 -> 2, 0x40, 8, "statevar"
+// test(uint): 0 -> 0, ""
+// test(uint): 42 -> 2, "statevar"
